XplatUIX11: Fixed arithmetic overflow when processing keystrokes with uptime of more than 25 days
In XKeyEvent, the time field is filled in with the time in milliseconds from the moment the X server is started.
This value has a 32-bit unsigned integer data type, and when processed in managed code
, it is converted to a 32-bit integer. This value is reset to 0 every 49.1 days and
starts over. An overflow will occur if the X server runs for more than 24 days without
restarting and the user presses a key in the WinForms application.

The original ClickStruct structure has been copied to the System.Windows.Forms.CarbonInternal/MouseHandler.cs
because Mac uses DateTime.Now.Ticks to fill it, which is of type Int64