XplatUIX11: Fixed double free or corruption (out)
Fixed the native error double free or corruption (out)

The error occurred due to a chain of calls
to region.GetHrgn (Graphics.FromHwnd(hwnd.whole_window)).
Probably a Graphics object obtained from Graphics.FromHwnd
could be released later or earlier than the Hwnd from which it was created.
In mono, this code looks like region.GetHrgn (null);
In the System.Drawing in mono returns null when passing
NativeObject, and in System.Drawing.Common is thrown away
ArgumentNullException. I tried not to create anything,
but to transfer it to wParam (IntPtr)1. There were no problems during the checks.