Fixed window hanging
After calling Destroywindow to handle the window and before receiving the event
DestroyNotify may issue the same handle in the XCreateWindow method,
and only then can DestroyNotify come for this handle.
A situation arises when a new window (control) is created with the handle of an existing
window. As a result, the Hwnd of the old one will be overwritten by the new one, and later upon receipt
DestroyNotify will destroy the new window, not the old one. Then we just hedge our bets,
if the new handle already has a window and it has the status of a zombie, then we boldly destroy it,
and we put a new one in the array. But we do the processing of DestroyNotify only if the
window has the zombie status.