XplatUIX11: The DestroyNotify event was ignored
The handle was removed from the array of handles using the Hwnd method.Dispose().
In conclusion, when extracting messages from the queue, it
was ignored, then that this handle was no longer
in the array and the hwnd= Hwnd.GetObjectFromWindow(xevent.AnyEvent.window) method
returned null. Checking that the handle is currently in the BeingDestroyed state
has been moved to the UpdateMessageQueue method before calling
Hwnd.GetObjectFromWindow(xevent.AnyEvent.window).